Corrected acceleration of enemies when PlayerShip changes direction

+++ Sprite.cpp
@@ -5,7 +5,7 @@
 #include "Sprite.h"
 #include "ST7735.h"
 
-void Sprite::Draw(int16_t x, uint8_t y){
+void Sprite::Draw(int x, uint8_t y){
   ST7735_DrawBitmap(y, x, image, w, h);
 }
